Move new autoscaler to builder package.
