+++ cluster-autoscaler/builder/autoscaler_factory.go
@@ -1,0 +1,174 @@
+/*
+Copyright The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package builder
+
+import (
+	"context"
+	"strings"
+
+	cqv1alpha1 "k8s.io/autoscaler/cluster-autoscaler/apis/capacityquota/autoscaling.x-k8s.io/v1alpha1"
+	cloudBuilder "k8s.io/autoscaler/cluster-autoscaler/cloudprovider/builder"
+	ca_context "k8s.io/autoscaler/cluster-autoscaler/context"
+	"k8s.io/autoscaler/cluster-autoscaler/core"
+	coreoptions "k8s.io/autoscaler/cluster-autoscaler/core/options"
+	"k8s.io/autoscaler/cluster-autoscaler/core/scaledown/pdb"
+	"k8s.io/autoscaler/cluster-autoscaler/core/utils"
+	"k8s.io/autoscaler/cluster-autoscaler/estimator"
+	"k8s.io/autoscaler/cluster-autoscaler/expander/factory"
+	ca_processors "k8s.io/autoscaler/cluster-autoscaler/processors"
+	"k8s.io/autoscaler/cluster-autoscaler/resourcequotas"
+	"k8s.io/autoscaler/cluster-autoscaler/resourcequotas/capacityquota"
+	"k8s.io/autoscaler/cluster-autoscaler/simulator/clustersnapshot/predicate"
+	"k8s.io/autoscaler/cluster-autoscaler/simulator/clustersnapshot/store"
+	csinodeprovider "k8s.io/autoscaler/cluster-autoscaler/simulator/csi/provider"
+	"k8s.io/autoscaler/cluster-autoscaler/simulator/drainability/rules"
+	draprovider "k8s.io/autoscaler/cluster-autoscaler/simulator/dynamicresources/provider"
+	"k8s.io/autoscaler/cluster-autoscaler/simulator/framework"
+	"k8s.io/autoscaler/cluster-autoscaler/utils/backoff"
+	"k8s.io/autoscaler/cluster-autoscaler/utils/errors"
+	"k8s.io/client-go/informers"
+)
+
+// NewAutoscaler creates an autoscaler of an appropriate type according to the parameters
+func NewAutoscaler(ctx context.Context, opts coreoptions.AutoscalerOptions, informerFactory informers.SharedInformerFactory) (core.Autoscaler, errors.AutoscalerError) {
+	err := initializeDefaultOptions(ctx, &opts, informerFactory)
+	if err != nil {
+		return nil, errors.ToAutoscalerError(errors.InternalError, err)
+	}
+	return core.NewStaticAutoscaler(
+		opts.AutoscalingOptions,
+		opts.FrameworkHandle,
+		opts.ClusterSnapshot,
+		opts.AutoscalingKubeClients,
+		opts.Processors,
+		opts.LoopStartObservers,
+		opts.CloudProvider,
+		opts.ExpanderStrategy,
+		opts.EstimatorBuilder,
+		opts.Backoff,
+		opts.ScaleUpFailuresRegistry,
+		opts.DebuggingSnapshotter,
+		opts.RemainingPdbTracker,
+		opts.ScaleUpOrchestrator,
+		opts.DeleteOptions,
+		opts.DrainabilityRules,
+		opts.DraProvider,
+		opts.QuotasTrackerOptions,
+		opts.MinQuotasTrackerOptions,
+		opts.CSIProvider,
+		opts.CapacityBufferPodsRegistry,
+	), nil
+}
+
+// Initialize default options if not provided.
+func initializeDefaultOptions(ctx context.Context, opts *coreoptions.AutoscalerOptions, informerFactory informers.SharedInformerFactory) error {
+	if opts.Processors == nil {
+		opts.Processors = ca_processors.DefaultProcessors(opts.AutoscalingOptions)
+	}
+	if opts.AutoscalingKubeClients == nil {
+		opts.AutoscalingKubeClients = ca_context.NewAutoscalingKubeClients(ctx, opts.AutoscalingOptions, opts.KubeClient, opts.InformerFactory)
+	}
+	if opts.FrameworkHandle == nil {
+		fwHandle, err := framework.NewHandle(ctx, opts.InformerFactory, opts.SchedulerConfig, opts.DynamicResourceAllocationEnabled, opts.CSINodeAwareSchedulingEnabled)
+		if err != nil {
+			return err
+		}
+		opts.FrameworkHandle = fwHandle
+	}
+	if opts.ClusterSnapshot == nil {
+		opts.ClusterSnapshot = predicate.NewPredicateSnapshot(store.NewBasicSnapshotStore(), opts.FrameworkHandle, opts.DynamicResourceAllocationEnabled, opts.PredicateParallelism, opts.CSINodeAwareSchedulingEnabled)
+	}
+	if opts.RemainingPdbTracker == nil {
+		opts.RemainingPdbTracker = pdb.NewBasicRemainingPdbTracker()
+	}
+	if opts.EstimatorBuilder == nil {
+		thresholds := []estimator.Threshold{
+			estimator.NewStaticThreshold(opts.MaxNodesPerScaleUp, opts.MaxNodeGroupBinpackingDuration),
+			estimator.NewSngCapacityThreshold(),
+			estimator.NewClusterCapacityThreshold(),
+		}
+		estimatorBuilder, err := estimator.NewEstimatorBuilder(
+			opts.EstimatorName,
+			estimator.NewThresholdBasedEstimationLimiter(thresholds),
+			estimator.NewDecreasingPodOrderer(),
+			/* EstimationAnalyserFunc */ nil,
+			opts.FastpathBinpackingEnabled,
+		)
+		if err != nil {
+			return err
+		}
+		opts.EstimatorBuilder = estimatorBuilder
+	}
+	if opts.Backoff == nil {
+		opts.Backoff =
+			backoff.NewIdBasedExponentialBackoff(opts.InitialNodeGroupBackoffDuration, opts.MaxNodeGroupBackoffDuration, opts.NodeGroupBackoffResetTimeout)
+	}
+	if opts.DrainabilityRules == nil {
+		opts.DrainabilityRules = rules.Default(opts.DeleteOptions)
+	}
+	if opts.DraProvider == nil && opts.DynamicResourceAllocationEnabled {
+		opts.DraProvider = draprovider.NewProviderFromInformers(informerFactory)
+	}
+	if opts.CloudProvider == nil {
+		opts.CloudProvider = cloudBuilder.NewCloudProvider(opts, informerFactory)
+	}
+
+	if opts.ExpanderStrategy == nil {
+		expanderFactory := factory.NewFactory()
+		expanderFactory.RegisterDefaultExpanders(opts.CloudProvider, opts.AutoscalingKubeClients, opts.KubeClient, opts.ConfigNamespace, opts.GRPCExpanderCert, opts.GRPCExpanderURL)
+		expanderStrategy, err := expanderFactory.Build(strings.Split(opts.ExpanderNames, ","))
+		if err != nil {
+			return err
+		}
+		opts.ExpanderStrategy = expanderStrategy
+	}
+	if opts.QuotasTrackerOptions.QuotaProvider == nil {
+		providers := []resourcequotas.Provider{resourcequotas.NewCloudQuotasProvider(opts.CloudProvider)}
+
+		if opts.CapacityQuotasEnabled {
+			// register informer here to disable lazy initialization
+			if _, err := opts.KubeCache.GetInformer(context.TODO(), &cqv1alpha1.CapacityQuota{}); err != nil {
+				return err
+			}
+			providers = append(providers, capacityquota.NewCapacityQuotasProvider(opts.KubeClientNew))
+		}
+		opts.QuotasTrackerOptions.QuotaProvider = resourcequotas.NewCombinedQuotasProvider(providers)
+	}
+	if opts.QuotasTrackerOptions.CustomResourcesProcessor == nil {
+		opts.QuotasTrackerOptions.CustomResourcesProcessor = opts.Processors.CustomResourcesProcessor
+	}
+	if opts.QuotasTrackerOptions.NodeFilter == nil {
+		virtualKubeletNodeFilter := utils.VirtualKubeletNodeFilter{}
+		opts.QuotasTrackerOptions.NodeFilter = resourcequotas.NewCombinedNodeFilter([]resourcequotas.NodeFilter{virtualKubeletNodeFilter})
+	}
+
+	if opts.MinQuotasTrackerOptions.QuotaProvider == nil {
+		opts.MinQuotasTrackerOptions.QuotaProvider = resourcequotas.NewCloudMinProvider(opts.CloudProvider)
+	}
+	if opts.MinQuotasTrackerOptions.CustomResourcesProcessor == nil {
+		opts.MinQuotasTrackerOptions.CustomResourcesProcessor = opts.Processors.CustomResourcesProcessor
+	}
+	if opts.MinQuotasTrackerOptions.NodeFilter == nil {
+		opts.MinQuotasTrackerOptions.NodeFilter = opts.QuotasTrackerOptions.NodeFilter
+	}
+
+	if opts.CSINodeAwareSchedulingEnabled {
+		csiProvider := csinodeprovider.NewCSINodeProviderFromInformers(informerFactory)
+		opts.CSIProvider = csiProvider
+	}
+	return nil
+}
